requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
mysql-connector-python==9.6.0
pdfkit==1.0.0
python-dotenv>=1.0.0
Werkzeug==3.1.5
gunicorn==21.2.0
resend==2.10.0
